third_party/blink/web_tests/fast/css/css-properties-values-api/at-property.html

<!DOCTYPE html>
<!--TODO(andruud): Migrate to WPT once @property is in the specification. -->
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="outer">
  <div id="target"></div>
</div>
<script>

let g_id = 0;

function generate_name() {
  g_id++;
  return `--property-${g_id}`;
}

function with_style_node(text, fn) {
  let node = document.createElement('style');
  node.textContent = text;
  try {
    document.body.append(node);
    fn(node);
  } finally {
    node.remove();
  }
}

function with_at_rule(desc, fn) {
  let name = typeof(desc.name) === 'undefined' ? generate_name() : desc.name;
  let text = `@property ${name} {`;
  if (typeof(desc.syntax) !== 'undefined')
    text += `syntax:${desc.syntax};`;
  if (typeof(desc.initialValue) !== 'undefined')
    text += `initial-value:${desc.initialValue};`;
  if (typeof(desc.inherits) !== 'undefined')
    text += `inherits:${desc.inherits};`;
  text += '}';
  with_style_node(text, (node) => fn(name, node.sheet.rules[0]));
}

function test_with_at_rule(desc, fn, description) {
  test(() => with_at_rule(desc, fn), description);
}

function test_with_style_node(text, fn, description) {
  test(() => with_style_node(text, fn), description);
}

// Parsing:

let uppercase_first = (x) => x.charAt(0).toUpperCase() + x.slice(1);
let to_camel_case = (x) => x.split('-')[0] + x.split('-').slice(1).map(uppercase_first).join('');

function test_parse_success(descriptor, value) {
  test_with_at_rule({ [to_camel_case(descriptor)]: value }, (name, rule) => {
    assert_equals(rule.style[to_camel_case(descriptor)], value);
  }, `Accepts valid value for '${descriptor}' descriptor [${value}]`);
}

function test_parse_failure(descriptor, value) {
  test_with_at_rule({ [to_camel_case(descriptor)]: value }, (name, rule) => {
    assert_equals(rule.style[to_camel_case(descriptor)], '');
  }, `Rejects invalid value for '${descriptor}' descriptor [${value}]`);
}

// syntax
test_parse_success('syntax', '"<color>"');
test_parse_success('syntax', '"<color> | none"');
test_parse_success('syntax', '"<color># | <image> | none"');
test_parse_success('syntax', '"foo | bar | baz"');
test_parse_success('syntax', '"*"');
test_parse_success('syntax', '"notasyntax"');

test_parse_failure('syntax', 'red');
test_parse_failure('syntax', 'rgb(255, 0, 0)');
test_parse_failure('syntax', '<color>');
test_parse_failure('syntax', 'foo | bar');

// initial-value
test_parse_success('initial-value', '10px');
test_parse_success('initial-value', 'rgb(1, 2, 3)');
test_parse_success('initial-value', 'red');
test_parse_success('initial-value', 'foo');
test_parse_success('initial-value', 'if(){}');
test_parse_success('initial-value', 'var(--x)');

// inherits
test_parse_success('inherits', 'true');
test_parse_success('inherits', 'false');

test_parse_failure('inherits', 'none');
test_parse_failure('inherits', '0');
test_parse_failure('inherits', '1');
test_parse_failure('inherits', '"true"');
test_parse_failure('inherits', '"false"');
test_parse_failure('inherits', 'calc(0)');

test_with_style_node('@property foo { }', (node) => {
  assert_equals(node.sheet.rules.length, 0);
}, 'Invalid property name does not parse [foo]');

test_with_style_node('@property -foo { }', (node) => {
  assert_equals(node.sheet.rules.length, 0);
}, 'Invalid property name does not parse [-foo]');

// Applying @property rules

function test_applied(syntax, initial, inherits, expected) {
  test_with_at_rule({
    syntax: `"${syntax}"`,
    initialValue: initial,
    inherits: inherits
  }, (name, rule) => {
    let actual = getComputedStyle(target).getPropertyValue(name);
    assert_equals(actual, expected);
  }, `Rule applied [${syntax}, ${initial}, ${inherits}]`);
}

function test_not_applied(syntax, initial, inherits) {
  test_with_at_rule({
    syntax: `"${syntax}"`,
    initialValue: initial,
    inherits: inherits
  }, (name, rule) => {
    let actual = getComputedStyle(target).getPropertyValue(name);
    assert_equals(actual, '');
  }, `Rule not applied [${syntax}, ${initial}, ${inherits}]`);
}

// syntax, initialValue, inherits, expected
test_applied('*', 'if(){}', false, 'if(){}');
test_applied('<angle>', '42deg', false, '42deg');
test_applied('<angle>', '1turn', false, '360deg');
test_applied('<color>', 'green', false, 'rgb(0, 128, 0)');
test_applied('<color>', 'rgb(1, 2, 3)', false, 'rgb(1, 2, 3)');
test_applied('<image>', 'url("http://a/")', false, 'url("http://a/")');
test_applied('<integer>', '5', false, '5');
test_applied('<length-percentage>', '10px', false, '10px');
test_applied('<length-percentage>', '10%', false, '10%');
test_applied('<length-percentage>', 'calc(10% + 10px)', false, 'calc(10% + 10px)');
test_applied('<length>', '10px', false, '10px');
test_applied('<number>', '2.5', false, '2.5');
test_applied('<percentage>', '10%', false, '10%');
test_applied('<resolution>', '50dppx', false, '50dppx');
test_applied('<resolution>', '96dpi', false, '1dppx');
test_applied('<time>', '10s', false, '10s');
test_applied('<time>', '1000ms', false, '1s');
test_applied('<transform-function>', 'rotateX(0deg)', false, 'rotateX(0deg)');
test_applied('<transform-list>', 'rotateX(0deg)', false, 'rotateX(0deg)');
test_applied('<transform-list>', 'rotateX(0deg) translateX(10px)', false, 'rotateX(0deg) translateX(10px)');
test_applied('<url>', 'url("http://a/")', false, 'url("http://a/")');

// inherits: true/false
test_applied('<color>', 'tomato', false, 'rgb(255, 99, 71)');
test_applied('<color>', 'tomato', true, 'rgb(255, 99, 71)');

test_with_at_rule({ syntax: '"*"', inherits: true }, (name, rule) => {
  try {
    outer.style.setProperty(name, 'foo');
    let actual = getComputedStyle(target).getPropertyValue(name);
    assert_equals(actual, 'foo');
  } finally {
    outer.style = '';
  }
}, 'Rule applied for "*", even with no initial value');

test_not_applied(undefined, 'green', false);
test_not_applied('<color>', undefined, false);
test_not_applied('<color>', 'green', undefined);
test_not_applied('<gandalf>', 'grey', false);
test_not_applied('gandalf', 'grey', false);
test_not_applied('<color>', 'notacolor', false);
test_not_applied('<length>', '10em', false);

// Inheritance

test_with_at_rule({
  syntax: '"<length>"',
  inherits: false,
  initialValue: '0px'
}, (name, rule) => {
  try {
    outer.style = `${name}: 40px`;
    assert_equals(getComputedStyle(outer).getPropertyValue(name), '40px');
    assert_equals(getComputedStyle(target).getPropertyValue(name), '0px');
  } finally {
    outer.style = '';
  }
}, 'Non-inherited properties do not inherit');

test_with_at_rule({
  syntax: '"<length>"',
  inherits: true,
  initialValue: '0px'
}, (name, rule) => {
  try {
    outer.style = `${name}: 40px`;
    assert_equals(getComputedStyle(outer).getPropertyValue(name), '40px');
    assert_equals(getComputedStyle(target).getPropertyValue(name), '40px');
  } finally {
    outer.style = '';
  }
}, 'Inherited properties inherit');

test_with_at_rule({
  syntax: '"<color>"',
  inherits: true,
  initialValue: 'green'
}, (name, rule) => {
  try {
    target.style = `--x:var(${name})`;
    assert_equals(getComputedStyle(target).getPropertyValue(name), 'rgb(0, 128, 0)');
  } finally {
    target.style = '';
  }
}, 'Initial values substituted as computed value');

</script>
